requirements.txt
praw
requests
opencv-python
numpy
python-dotenv
